split projection code off for testing
