+++ semantic_inference_ros/src/backprojection_nodelet.cpp
@@ -1,121 +1,35 @@
 #include <config_utilities/config_utilities.h>
 #include <config_utilities/parsing/commandline.h>
-// #include <config_utilities/types/path.h>
+#include <config_utilities/types/path.h>
 #include <glog/logging.h>
 #include <ianvs/image_subscription.h>
 #include <message_filters/subscriber.h>
 #include <message_filters/sync_policies/approximate_time.h>
 #include <message_filters/synchronizer.h>
-#include <semantic_inference/image_recolor.h>
 #include <tf2_ros/transform_listener.h>
 
 #include <cstdint>
 #include <string>
 
 #include <cv_bridge/cv_bridge.hpp>
-#include <image_geometry/pinhole_camera_model.hpp>
-#include <opencv2/core.hpp>
 #include <rclcpp/node.hpp>
-#include <sensor_msgs/msg/camera_info.hpp>
-#include <sensor_msgs/msg/image.hpp>
-#include <sensor_msgs/msg/point_cloud2.hpp>
-#include <sensor_msgs/point_cloud2_iterator.hpp>
 #include <tf2_eigen/tf2_eigen.hpp>
 
+#include "semantic_inference_ros/pointcloud_projection.h"
 #include "semantic_inference_ros/ros_log_sink.h"
-#include <Eigen/Geometry>
 
 namespace semantic_inference {
 
-using message_filters::Synchronizer;
+using message_filters::sync_policies::ApproximateTime;
 using sensor_msgs::msg::CameraInfo;
 using sensor_msgs::msg::Image;
 using sensor_msgs::msg::PointCloud2;
 using sensor_msgs::msg::PointField;
 
-struct ProjectionConfig {
-  bool use_lidar_frame = true;
-  bool discard_out_of_view = false;
-  int16_t unknown_label = 0;
-};
-
-void declare_config(ProjectionConfig& config) {
-  using namespace config;
-  name("ProjectionConfig::Config");
-  field(config.use_lidar_frame, "use_lidar_frame");
-  field(config.discard_out_of_view, "discard_out_of_view");
-  field(config.unknown_label, "unknown_label");
-}
-
-void projectSemanticImage(const ProjectionConfig& config,
-                          const CameraInfo& intrinsics,
-                          const cv::Mat& image,
-                          const PointCloud2& cloud,
-                          const Eigen::Isometry3f& image_T_cloud,
-                          PointCloud2& output) {
-  image_geometry::PinholeCameraModel model;
-  model.fromCameraInfo(intrinsics);
-
-  sensor_msgs::PointCloud2Modifier mod(output);
-  // clang-format off
-  mod.setPointCloud2Fields(4,
-                           "x", 1, PointField::FLOAT32,
-                           "y", 1, PointField::FLOAT32,
-                           "z", 1, PointField::FLOAT32,
-                           "label", 1, PointField::INT32);
-  // clang-format on
-  mod.resize(cloud.width, cloud.height);
-
-  // techinically we could probably just have one iterator, but (small) chance that
-  // someone sends a non-contiguous pointcloud
-  auto x_in = sensor_msgs::PointCloud2ConstIterator<float>(cloud, "x");
-  auto y_in = sensor_msgs::PointCloud2ConstIterator<float>(cloud, "y");
-  auto z_in = sensor_msgs::PointCloud2ConstIterator<float>(cloud, "z");
-  auto label_iter = sensor_msgs::PointCloud2Iterator<int32_t>(output, "label");
-  while (x_in != x_in.end()) {
-    const Eigen::Vector3f p_cloud(*x_in, *y_in, *z_in);
-    const Eigen::Vector3f p_image = image_T_cloud * p_cloud;
-    ++x_in;
-    ++y_in;
-    ++z_in;
-
-    int u = -1;
-    int v = -1;
-    if (p_image.z() > 0.0f) {
-      const auto& pixel =
-          model.project3dToPixel(cv::Point3d(p_image.x(), p_image.y(), p_image.z()));
-      u = std::round(pixel.x);
-      v = std::round(pixel.y);
-    }
-
-    const auto in_view = u >= 0 && u < image.cols && v >= 0 && v < image.rows;
-    *label_iter = in_view ? image.at<int32_t>(v, u) : config.unknown_label;
-    ++label_iter;
-
-    if (!in_view && config.discard_out_of_view) {
-    }
-
-    // TODO(nathan) optionally null out out-of-view points
-  }
-}
-
-void colorPointcloud(const ImageRecolor& recolor, PointCloud2& input) {
-  auto label_iter = sensor_msgs::PointCloud2ConstIterator<int32_t>(input, "label");
-  auto color_iter = sensor_msgs::PointCloud2Iterator<float>(input, "rgb");
-  while (label_iter != label_iter.end()) {
-    const auto& color = recolor.getColor(*label_iter);
-    color_iter[0] = color[0];
-    color_iter[1] = color[1];
-    color_iter[2] = color[2];
-    ++label_iter;
-    ++color_iter;
-  }
-}
-
 struct BackprojectionNode : public rclcpp::Node {
  public:
-  using SyncPolicy =
-      message_filters::sync_policies::ApproximateTime<Image, CameraInfo, PointCloud2>;
+  using SyncPolicy = ApproximateTime<Image, CameraInfo, PointCloud2>;
+  using Sync = message_filters::Synchronizer<SyncPolicy>;
 
   struct Config {
     ProjectionConfig projection;
@@ -138,7 +52,7 @@ struct BackprojectionNode : public rclcpp::Node {
   ianvs::ImageSubscription image_sub_;
   message_filters::Subscriber<CameraInfo> info_sub_;
   message_filters::Subscriber<PointCloud2> cloud_sub_;
-  std::unique_ptr<message_filters::Synchronizer<SyncPolicy>> sync;
+  std::unique_ptr<Sync> sync;
 
   tf2_ros::Buffer tf_buffer_;
   tf2_ros::TransformListener tf_listener_;
@@ -178,9 +92,7 @@ BackprojectionNode::BackprojectionNode(const rclcpp::NodeOptions& options)
   image_sub_.subscribe("image", config.input_queue_size);
   info_sub_.subscribe(this, "camera_info", qos.get_rmw_qos_profile());
   cloud_sub_.subscribe(this, "cloud", qos.get_rmw_qos_profile());
-
-  sync.reset(
-      new Synchronizer<SyncPolicy>(SyncPolicy(10), image_sub_, info_sub_, cloud_sub_));
+  sync = std::make_unique<Sync>(SyncPolicy(10), image_sub_, info_sub_, cloud_sub_);
   sync->registerCallback(std::bind(&BackprojectionNode::callback, this, _1, _2, _3));
 }
